pyproject.toml
[project]
name = "url-shortener-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beanie>=1.29.0",
    "fastapi[standard]>=0.115.11",
    "granian>=2.1.2",
    "jinja2>=3.1.6",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
